fixing some bugs, separating tests
