--- tests/feature/basic.test.js
@@ -1,3 +1,0 @@
-test('Jest is up and running . . . ', () => {
-    expect(2 + 2).toEqual(4)
-})
